+++ graylog2-web-interface/templates/index.html.template
@@ -8,6 +8,7 @@
     <script defer src="/<%= vendorModule().files.js[vendorChunk] %>"></script>
     <% } %>
     <%= htmlWebpackPlugin.tags.headTags %>
+    <script defer src="/plugin.org.graylog.plugins.enterprise.EnterprisePlugin.js"></script>
   </head>
   <body>
     <%= htmlWebpackPlugin.tags.bodyTags %>
