+++ graylog2-web-interface/webpack.bundled.js
@@ -17,4 +17,4 @@
 const webpackConfig = require('./webpack.combined.config');
 const vendorConfig = require('./webpack.vendor');
 
-module.exports = [vendorConfig, webpackConfig];
+module.exports = [vendorConfig, ...webpackConfig];
