+++ graylog2-web-interface/webpack.combined.config.js
@@ -18,8 +18,6 @@ const fs = require('fs');
 const path = require('path');
 
 const { globSync } = require('glob');
-const merge = require('webpack-merge');
-const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 
 const TARGET = process.env.npm_lifecycle_event;
 const WEB_MODULES = path.resolve(__dirname, './web-modules.json');
@@ -53,13 +51,13 @@ const configsFromGlob = () => {
 };
 
 // eslint-disable-next-line no-nested-ternary
-const pluginConfigs = process.env.disable_plugins === 'true'
+const pluginConfigFiles = process.env.disable_plugins === 'true'
   ? []
   : fs.existsSync(WEB_MODULES)
     ? configsFromWebModule(WEB_MODULES)
     : configsFromGlob();
 
-if (pluginConfigs.some((config) => config.includes('graylog-plugin-cloud/server-plugin'))) {
+if (pluginConfigFiles.some((config) => config.includes('graylog-plugin-cloud/server-plugin'))) {
   process.env.IS_CLOUD = true;
 }
 
@@ -67,41 +65,26 @@ process.env.web_src_path = path.resolve(__dirname);
 
 // eslint-disable-next-line import/no-dynamic-require
 const webpackConfig = require(path.resolve(__dirname, './webpack.config.js'));
+// eslint-disable-next-line import/no-dynamic-require,global-require
+const pluginConfigs = pluginConfigFiles.map((file) => require(file));
 
-const mergedPluginConfigs = pluginConfigs
-  // eslint-disable-next-line global-require,import/no-dynamic-require
-  .map((configFile) => require(configFile))
-  .reduce((config, pluginConfig) => merge.smart(config, pluginConfig), {});
-
-const mergedWebpackConfig = (TARGET === 'start')
-  ? merge(webpackConfig, {
-    plugins: [
-      new ForkTsCheckerWebpackPlugin(),
-    ],
-  })
-  : webpackConfig;
-
-const finalConfig = merge.smart(mergedPluginConfigs, mergedWebpackConfig);
+const allConfigs = [webpackConfig, ...pluginConfigs];
 
 // We need to inject webpack-hot-middleware to all entries, ensuring the app is able to reload on changes.
 if (TARGET === 'start') {
-  const hmrEntries = {};
   const webpackHotMiddlewareEntry = 'webpack-hot-middleware/client?reload=true';
 
-  Object.keys(finalConfig.entry).forEach((entryKey) => {
-    const entryValue = finalConfig.entry[entryKey];
-    const hmrValue = [webpackHotMiddlewareEntry];
+  allConfigs.forEach((finalConfig) => {
+    const hmrEntries = {};
 
-    if (Array.isArray(entryValue)) {
-      hmrValue.push(...entryValue);
-    } else {
-      hmrValue.push(entryValue);
-    }
+    Object.keys(finalConfig.entry).forEach((entryKey) => {
+      const entryValue = finalConfig.entry[entryKey];
+      hmrEntries[entryKey] = [webpackHotMiddlewareEntry].concat(Array.isArray(entryValue) ? entryValue : [entryValue]);
+    });
 
-    hmrEntries[entryKey] = hmrValue;
+    // eslint-disable-next-line no-param-reassign
+    finalConfig.entry = hmrEntries;
   });
-
-  finalConfig.entry = hmrEntries;
 }
 
-module.exports = finalConfig;
+module.exports = allConfigs;
